Make tests run properly on windows
Builds on windows append '.exe' to built files. This means test assertions and
the 'buildRetool' step need to be slightly different on windows and nonwindows.
